alamofire classの作成
